+++ src/Constants.js
@@ -1,5 +1,5 @@
-// const apiUrl = "http://localhost:5000"
-const apiUrl = "https://api.prijslijst.info";
+const apiUrl = "http://localhost:5000";
+// const apiUrl = "https://api.prijslijst.info";
 
 export const imageUrl = "https://d3sticxdmgvhkp.cloudfront.net/";